reorg RAI test
