+++ app_target.py
@@ -2,12 +2,14 @@ import requests
 from typing_extensions import Self
 from typing import TypedDict
 from promptflow.tracing import trace
+from dotenv import load_dotenv
 
+load_dotenv()
 
 class ModelEndpoints:
-    def __init__(self: Self, env: dict, model_type: str) -> str:
-        self.env = env
-        self.model_type = model_type
+    def __init__() -> str:
+        self.model_endpoint = os.environ["MODEL_ENDPOINT"]
+        self.model_key = os.environ["MODEL_KEY"]
 
     class Response(TypedDict):
         query: str
@@ -15,25 +17,21 @@ class ModelEndpoints:
 
     @trace
     def __call__(self: Self, question: str) -> Response:
-        #self.model_type == "onnx-model":
-        output = self.call_onnx_endpoint(question)
+        # We can integrate with any model endpoint
+        output = self.call_model_endpoint(question)
         return output
 
     def query(self: Self, endpoint: str, headers: str, payload: str) -> str:
         response = requests.post(url=endpoint, headers=headers, json=payload)
         return response.json()
 
-    def call_onnx_endpoint(self: Self, question: str) -> Response:
-        endpoint = self.env["onnx-model"]["endpoint"]
-        key = self.env["onnx-model"]["key"]
+    def call_model_endpoint(self: Self, question: str) -> Response:
+        endpoint = self.model_endpoint
+        key = self.model_key
 
         headers = {"Content-Type": "application/json", "api-key": key}
-        #headers = {"Content-Type": "application/json"}
-
-        #payload = {"messages": [{"role": "user", "content": question}], "max_tokens": 500}
-        payload = {"text": question}
-        
+        payload = {"text": question}        
         output = self.query(endpoint=endpoint, headers=headers, payload=payload)
-        #answer = output["choices"][0]["message"]["content"]
         answer = output["response"]
+
         return {"query": question, "response": answer}
